IDE: Move ReplaceInfixOrOperatorCallFix to idea-frontend-independent.
